Move auth controllers to routes dir
